test(native): co-locate token and trading atom tests
